pyproject.toml
[project]
name = "cardstore"
version = "0.1.0"
description = "Card Store - 开源虚拟物品交易平台"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "tortoise-orm[asyncpg]>=0.23.0",
    "aerich>=0.8.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "aiosmtplib>=3.0.0",
    "email-validator>=2.2.0",
    "loguru>=0.7.0",
    "aiofiles>=25.1.0",
    "httpx>=0.28.1",
    "redis>=7.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
